refactor navigation filtering
