+++ core-ui/src/state/navigation/filterByScope.ts
@@ -1,0 +1,24 @@
+import { NavNode } from '../types';
+
+export const filterByScope = (
+  navList: NavNode[],
+  scope: 'cluster' | 'namespace',
+) => {
+  if (!(scope === 'cluster' || scope === 'namespace')) {
+    console.error('Navigation scope is not defined.');
+    return [];
+  }
+  const isNamespace = scope === 'namespace';
+  const isCluster = scope === 'cluster';
+
+  const filteredList = navList.filter(resource => {
+    if (isNamespace) {
+      return resource.namespaced;
+    }
+    if (isCluster) {
+      return !resource.namespaced;
+    }
+    return false;
+  });
+  return filteredList;
+};
